Add encoding cred offer data into base64
